Tidy up code and add tests. (#1)
* Take other snakes bodies into account

* Tidy up

* Silly mistake

* Added logging

* Updated for loops to use correct index

* If all else fails choose a random valid move

* Don't go off the board please

* Custom Looks

* Small refactor and added first chase tail.

* Make snek get big first

* Silly logic mistake.

Really need to begin to add tests here.

* Silly syntax error.

* Small tweaks.

* Added first test suite.

First pass at running suite of cases on a single unit. Works fairly well
but was a lot of boiler plate.

Hopefully this can be abstracted/reduced into something similar.

* Added more tests for the current behaviours.

Ensured behaviours returned expected coordinates.

* Removed extra log.

* Log coord to move to.

* Add game state logging back

* Add some more logging.

* Remove unused imports

* Return random valid move if all else fails.

If we cannot move in the directions we "want" to go in then at least
choose a valid move.